sf::vector2i

+++ The platformer/Engine/GameObject.h
@@ -8,7 +8,7 @@
 struct GameObjectProperties
 {
 public:
-	GameObjectProperties(const std::string& textureId, const sf::Vector2f& position, const int& width, const int& height)
+	GameObjectProperties(const std::string& textureId, const sf::Vector2i& position, const int& width, const int& height)
 	{
 		m_textureId = textureId;
 		m_position.x = position.x;
@@ -20,7 +20,7 @@ public:
 	std::string m_textureId;
 	int m_width;
 	int m_height;
-	sf::Vector2f m_position;
+	sf::Vector2i m_position;
 };
 
 class GameObject
@@ -33,7 +33,7 @@ public:
 		m_width = properties.m_width;
 		m_height = properties.m_height;
 
-		m_position = sf::Vector2f(properties.m_position.x, properties.m_position.y);
+		m_position = sf::Vector2i(properties.m_position.x, properties.m_position.y);
 		float pointX = properties.m_position.x + (float)properties.m_width / 2.f;
 		float pointY = properties.m_position.x + (float)properties.m_height / 2.f;
 		m_origin = Point(pointX, pointY);
@@ -53,7 +53,7 @@ public:
 	virtual void Render() = 0;
 	virtual void Update(float dt) = 0;
 protected:
-	sf::Vector2f m_position;
+	sf::Vector2i m_position;
 	Point m_origin;
 	Collider m_collider;
 	GameObjectProperties m_properties;
